Hardening Across Memory, Concurrency, and Recording (#121)
Harden C64Stream across all 34 audit findings

Complete the C64STR-001 through C64STR-034 hardening audit, covering
memory safety, concurrency, recording, device control, multi-instance
behaviour, filesystem handling, and platform compatibility.

Lifecycle and concurrency:

* Prevent retry workers from starting during source destruction and snapshot
  REST configuration coherently under lock (C64STR-001, C64STR-003).
* Publish frame dimensions as a coherent pair and maintain independent,
  thread-safe palette LUTs for each source (C64STR-008, C64STR-014).
* Serialize CSV recording lifecycle and writes, and eliminate the
  jitter-buffer torn-slot race during concurrent push, pop, and delay changes
  (C64STR-018, C64STR-024, C64STR-026).
* Assign distinct default UDP port pairs to multiple source instances
  (C64STR-016).

REST and device control:

* Retarget existing REST transports for password changes and SWITCH_DEVICE,
  allowing scripts to continue on the new device without rebuilding their
  cached transport objects (C64STR-002, C64STR-017).
* Dispatch joystick, menu, reset, and reboot commands through an asynchronous
  FIFO instead of blocking the OBS UI thread (C64STR-022).
* Enforce REST response-size limits after every curl reset and reset
  audio/video synchronisation state safely (C64STR-019, C64STR-020).

Recording:

* Report recording open failures correctly and clean up failed frame writes
  without leaving corrupt or dangling recording state (C64STR-009,
  C64STR-032).
* Write accurate WAV sample rates and clamp legacy 32-bit size fields instead
  of allowing them to wrap (C64STR-010, C64STR-011).
* Roll AVI recordings safely on size or geometry changes, generate
  deterministic continuation names, produce independently valid segment
  headers, and avoid per-frame header rewrites and flushes (C64STR-012,
  C64STR-013, C64STR-034).

Input and memory safety:

* Validate malformed network, script, and effect input before it reaches
  fixed-size buffers, allocations, or rendering paths (C64STR-007,
  C64STR-021, C64STR-023, C64STR-027, C64STR-028, C64STR-029,
  C64STR-030).
* Clamp packet-derived frame heights to valid PAL/NTSC bounds and restrict
  logo copies to the logo buffer's actual dimensions.
* Enforce finite script execution and recursion limits (C64STR-031).

Filesystem and platform handling:

* Replace shell-based directory creation, detect every short file write, and
  harden path and DNS handling (C64STR-004, C64STR-005, C64STR-006,
  C64STR-033).
* Prevent playlist scans from looping through symlink cycles or exceeding
  their recursion-depth limit (C64STR-015).
* Correctly handle macOS special keys without interpreting unresolved Carbon
  keycodes as ASCII or Windows virtual-key values (C64STR-025).

Add direct regression coverage for every audit finding, including concurrent
TSan tests, ASan memory-safety tests, recording-format validation, REST mock
tests, and multi-instance scenarios. The full test suite passes under the
supported sanitizer configurations, Windows test builds and macOS warning
gates are clean, and two-device SWITCH_DEVICE continuation has been verified
against real C64 Ultimate and Ultimate 64 hardware.

+++ src/record/c64-record-audio.h
@@ -20,7 +20,7 @@ struct c64_source;
 
 // Audio recording functions (WAV format)
 void c64_audio_write_wav_header(FILE *file, uint32_t sample_rate, uint16_t channels, uint16_t bits_per_sample);
-void c64_audio_finalize_wav_header(FILE *file, uint32_t data_size);
+void c64_audio_finalize_wav_header(FILE *file, uint64_t data_size);
 void c64_audio_record_data(struct c64_source *context, const uint8_t *audio_data, size_t data_size);
 
 #endif // C64_RECORD_AUDIO_H
